more refactoring to decouple things. Movie select logic still broken.
